Add files via upload
This project gives a view of how negative pricing effects the business of electricity trading and how factors such as 'Demand Forecast' in MWh and 'Wind Forecast' in MWh are correlated with the Pricing (Euros/hour). One of the solutions to the uncertainty of Wind availability and Demand is installation of efficient Battery storage systems.